Writing part of the streaming rework
This may not be a real enhancement since it seems the 3 supported
drivers (SQLite, MySQL and PostgreSQL) load the entire file content
before writing anyway. Thus, waiting for the PDO/LOB drivers to
be optimized, this patch could be seen as a duplication of code
without immediate and practical benefits.